Add full withdraw/bottle functionality

+++ src/main/java/me/colingrimes/nightbank/config/Settings.java
@@ -12,13 +12,13 @@ import static me.colingrimes.midnight.config.option.OptionFactory.option;
 public interface Settings {
 
 	// Banknote configurations.
-	Option<Long> WITHDRAW_MIN_AMOUNT = option("withdraw.min-amount", 1L);
-	Option<Long> WITHDRAW_MAX_AMOUNT = option("withdraw.max-amount", 1000000L);
+	Option<Double> WITHDRAW_MIN_AMOUNT = option("withdraw.min-amount", 1D);
+	Option<Double> WITHDRAW_MAX_AMOUNT = option("withdraw.max-amount", 1000000D);
 	Option<ItemStack> WITHDRAW_ITEM = option("withdraw", Items.of(Material.PAPER).name("&cInvalid Configuration.").build());
 
 	// Bottle configurations.
-	Option<Long> BOTTLE_MIN_AMOUNT = option("bottle.min-amount", 1L);
-	Option<Long> BOTTLE_MAX_AMOUNT = option("bottle.max-amount", 1000000L);
+	Option<Integer> BOTTLE_MIN_AMOUNT = option("bottle.min-amount", 1);
+	Option<Integer> BOTTLE_MAX_AMOUNT = option("bottle.max-amount", 1000000);
 	Option<ItemStack> BOTTLE_ITEM = option("bottle", Items.of(Material.EXPERIENCE_BOTTLE).name("&cInvalid Configuration.").build());
 
 	// Other configurations.
